MultipleChanges
Added the screen is centered when it gets to the game screen. Sharks get
removed when the hit the far left at -300(due to shark size) in
preparation for adding in lives. 
We need to add:
Lives and Failscreen(which brings you back to main menu)
Score and Score saves/display
Multiple modes(switching what the background for the Game.fxml is) the
switching of sharks is already made easy.
The toggleable sound button
Pause button and a pause state (might have to drop this idea if we can't
get it fast enough)